envoie de mail

+++ src/main/java/org/sid/dao/Type_ServiceRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-public interface Type_ServiceRepository extends JpaRepository<TypeService, Long> {
+public interface Type_ServiceRepository extends JpaRepository<TypeService, Integer> {
 }
